requirements.txt
discord.py==1.2.3
motor==2.0.0
dnspython==1.16.0
aiohttp==3.5.4
python-dotenv==0.10.3
sanic==19.6.0
